requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
azure-core==1.15.0
azure-storage-blob==12.8.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
fastapi==0.65.2
h11==0.12.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
multidict==5.1.0
oauthlib==3.1.1
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
yarl==1.6.3
